refactor(web-routes): organize library routes into subfolder
